% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.R
\name{test_claims}
\alias{test_claims}
\title{Collect and summarize claim tests}
\usage{
test_claims(...)
}
\arguments{
\item{...}{\code{claim_test} objects, or lists containing them.}
}
\value{
A \code{claim_test_summary} object.
}
\description{
Collect and summarize claim tests
}
